+++ src/dkplsqda.jl
@@ -41,8 +41,7 @@ function dkplsqda(X, y, weights::Weight; kwargs...)
     @inbounds for i = 1:par.nlv
         fitm_da[i] = qda(vcol(fitm_emb.T, 1:i), y, weights; kwargs...)
     end
-    fitm = (fitm_emb = fitm_emb, fitm_da = fitm_da)
-    Plsprobda(fitm, res.lev, ni, par)
+    Plsprobda(fitm_emb, fitm_da, res.lev, ni, par) 
 end