+++ src/kplskdeda.jl
@@ -42,8 +42,7 @@ function kplskdeda(X, y, weights::Weight; kwargs...)
     @inbounds for a = 1:par.nlv
         fitm_da[a] = kdeda(vcol(fitm_emb.T, 1:a), y; kwargs...)
     end
-    fitm = (fitm_emb = fitm_emb, fitm_da = fitm_da)
-    Plsprobda(fitm, res.lev, ni, par)
+    Plsprobda(fitm_emb, fitm_da, res.lev, ni, par) 
 end