+++ src/mbplslda.jl
@@ -117,7 +117,7 @@ Compute latent variables (LVs; = scores) from
 * `nlv` : Nb. LVs to consider.
 """ 
 function transf(object::Mbplsprobda, Xbl; nlv = nothing)
-    transf(object.fitm.fitm_emb, Xbl; nlv)
+    transf(object.fitm_emb, Xbl; nlv)
 end
 
 """
@@ -132,15 +132,15 @@ function predict(object::Mbplsprobda, Xbl; nlv = nothing)
     Q = eltype(Xbl[1][1, 1])
     Qy = eltype(object.lev)
     m = nro(Xbl[1])
-    a = size(object.fitm.fitm_emb.T, 2)
+    a = size(object.fitm_emb.T, 2)
     isnothing(nlv) ? nlv = a : nlv = min(a, minimum(nlv)):min(a, maximum(nlv))
     le_nlv = length(nlv)
     pred = list(Matrix{Qy}, le_nlv)
     posterior = list(Matrix{Q}, le_nlv)
     @inbounds for i in eachindex(nlv)
         znlv = nlv[i]
-        T = transf(object.fitm.fitm_emb, Xbl; nlv = znlv)
-        zres = predict(object.fitm.fitm_da[znlv], T)
+        T = transf(object.fitm_emb, Xbl; nlv = znlv)
+        zres = predict(object.fitm_da[znlv], T)
         z =  mapslices(argmax, zres.posterior; dims = 2) 
         pred[i] = reshape(recod_indbylev(z, object.lev), m, 1)
         posterior[i] = zres.posterior