+++ src/plskdeda.jl
@@ -48,7 +48,7 @@ fit!(model, Xtrain, ytrain)
 fitm.lev
 fitm.ni
 
-fitm_emb = fitm.fitm.fitm_emb ;
+fitm_emb = fitm.fitm_emb ;
 @head fitm_emb.T
 @head transf(model, Xtrain)
 @head transf(model, Xtest)
@@ -86,8 +86,7 @@ function plskdeda(X, y, weights::Weight; kwargs...)
     @inbounds for i = 1:par.nlv
         fitm_da[i] = kdeda(vcol(fitm_emb.T, 1:i), y; kwargs...)
     end
-    fitm = (fitm_emb = fitm_emb, fitm_da = fitm_da)
-    Plsprobda(fitm, res.lev, ni, par)
+    Plsprobda(fitm_emb, fitm_da, res.lev, ni, par) 
 end